refactor(api): split dependency cache helpers
